# swagger.json
{"openapi":"3.0.0","paths":{"/auth/signup":{"post":{"operationId":"AuthenticationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuthenticationDto"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"account successfully created"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Auth"]}},"/auth/account/verification":{"post":{"operationId":"AuthenticationController_verifyEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpDto"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"account successfully created"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthenticationController_findOne","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"account login successfully"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Auth"]}},"/auth/profile":{"get":{"operationId":"AuthenticationController_userProfile","parameters":[],"responses":{"200":{"description":"successful"},"201":{"description":"account login successfully"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/products/create":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"product successfully created"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Products"],"security":[{"bearer":[]}]}},"/products":{"get":{"operationId":"ProductsController_findAll","parameters":[],"responses":{"200":{"description":"successful"},"201":{"description":"retrieve products successfully"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Products"],"security":[{"bearer":[]}]}},"/products/{productId}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"productId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"successful"},"201":{"description":"product retrieve successfully"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Products"],"security":[{"bearer":[]}]},"patch":{"operationId":"ProductsController_update","parameters":[{"name":"productId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"update product successfully"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Products"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"productId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"successful"},"201":{"description":"product successful removed"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Products"],"security":[{"bearer":[]}]}}},"info":{"title":"SWITCHIVE ASSESSMENT","description":"API Descriptions","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateAuthenticationDto":{"type":"object","properties":{"email":{"type":"string","example":"switchive@gmail.com"},"first_name":{"type":"string","example":"Switch"},"last_name":{"type":"string","example":"Chive"},"password":{"type":"string","example":"Chive@123"}},"required":["email","first_name","last_name","password"]},"OtpDto":{"type":"object","properties":{"email":{"type":"string","example":"vamp@gmail.com"},"otp":{"type":"string","example":"8790"}},"required":["email","otp"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string","description":"user unique email","example":"switchive@gmail.com"},"password":{"type":"string","example":"Swtich@123"}},"required":["email","password"]},"CreateProductDto":{"type":"object","properties":{"product_name":{"type":"string","example":"switchive@gmail.com"},"product_description":{"type":"string","example":"Description of product name"},"product_price":{"type":"number","example":5504},"product_image_url":{"type":"string","example":"https://upload.wikimedia.org/wikipedia/commons/thumb/4/46/Bitcoin.svg/1200px-Bitcoin.svg.png"}},"required":["product_name","product_description","product_price","product_image_url"]},"UpdateProductDto":{"type":"object","properties":{"product_name":{"type":"string","example":"switchive@gmail.com"},"product_description":{"type":"string","example":"Description of product name"},"product_price":{"type":"number","example":5504},"product_image_url":{"type":"string","example":"https://upload.wikimedia.org/wikipedia/commons/thumb/4/46/Bitcoin.svg/1200px-Bitcoin.svg.png"}}}}}}